[KAN-210] fix(dashboard): 샘플 데이터 생성 로직 수정
